requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.26.4
faiss-cpu==1.8.0
ffmpeg-python==0.2.0
filelock==3.12.2
ftfy==6.2.0
future==1.0.0
idna==3.4
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
natsort==8.4.0
networkx==3.1
numpy==1.25.1
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
Pillow==10.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.0.3
wcwidth==0.2.13
